refactor: Add a producer package for separation of concerns
- Extract (Reader) Transform (Processor) Load (Producer).
